Moved CheckResult filtering into custom hook
